Added functionality to buttons in modal
- Item.js:
Generalized an addItem(qtt) function from the temporary one in onClick.
Added a state to keep track of quantity inside a modal. That state variable is passed as argument to addItem() when the Save Changes button is pressed, meaning the quantity will only be reflected in the cart if user chooses to save changes.
Inside a modal, quantity state variable cannot be set below 1 otherwise it doesn't interact well with items already in cart (cart with a negative quantity item).

- App.css:
Removed the unused styling.